% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_prioritize.R
\name{benefit}
\alias{benefit}
\title{Calculate taxon conservation benefit}
\usage{
benefit(x, lambda = 1)
}
\arguments{
\item{x}{Fraction of taxon range protected (value between 0 and 1).}

\item{lambda}{Shape parameter.}
}
\value{
Value between 0 and 1.
}
\description{
Nonlinear function that converts proportion of range conserved into conservation "benefit."
}
