% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylospatial.R
\name{build_tree_ranges}
\alias{build_tree_ranges}
\title{Derive ranges for the internal nodes of a tree}
\usage{
build_tree_ranges(tree, tip_comm, fun = NULL)
}
\arguments{
\item{tree}{Phylogeny (object of class "phylo").}

\item{tip_comm}{Site-by-taxon community matrix for terminal taxa.}

\item{fun}{Function used to calculate the community quantity for a clade based on the community quantities
for the set of tips in the clade that are present in a given site.}
}
\value{
A site-by-taxon matrix with a column for every branch, including terminals and internal edges.
}
\description{
Derive ranges for the internal nodes of a tree
}
