% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_tree_ranges_fast}
\alias{build_tree_ranges_fast}
\title{Fast clade range building with precomputed descendants}
\usage{
build_tree_ranges_fast(
  tree,
  tip_comm,
  data_type,
  descendants,
  clade_fun = NULL
)
}
\arguments{
\item{tree}{A phylo object}

\item{tip_comm}{Tip community matrix (sites x tips)}

\item{data_type}{Data type string}

\item{descendants}{Precomputed list from precompute_descendants()}

\item{clade_fun}{Custom aggregation function (used if data_type is "other")}
}
\value{
Full community matrix (sites x edges)
}
\description{
Lightweight version of build_tree_ranges() that uses precomputed
descendant information and skips validation.
}
\keyword{internal}
