% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tip_comm}
\alias{get_tip_comm}
\title{Get community data for terminal taxa}
\usage{
get_tip_comm(ps, spatial = F)
}
\arguments{
\item{ps}{Object of class \code{phylospatial}.}

\item{spatial}{Logical indicating whether a spatial (\code{SpatRaster} or \code{sf}) object should be returned. If FALSE (the default), a
matrix is returned.}
}
\value{
A site-by-taxon matrix with a column for every terminal branch.
}
\description{
The community matrix contained in a \code{phylospatial} object includes a column for every terminal and internal branch.
This function returns the sub-matrix that includes only the terminal taxa.
}
