% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_dissim.R
\name{ps_add_dissim}
\alias{ps_add_dissim}
\title{Add community dissimilarity data to a \code{phylospatial} object}
\usage{
ps_add_dissim(ps, method = "bray", endemism = FALSE, normalize = TRUE, ...)
}
\arguments{
\item{ps}{phylospatial object.}

\item{method}{Character indicating the dissimilarity index to use, passed to \link[vegan]{vegdist}. The default is
"bray", i.e. Bray-Curtis distance, also known as quantitative Sorensen's. See \link[vegan]{vegdist} for a complete
list of options.}

\item{endemism}{Logical indicating whether community values should be divided by column totals (taxon range sizes)
to derive endemism.}

\item{normalize}{Logical indicating whether community values should be divided by row totals (community sums). If \code{TRUE},
dissimilarity is based on proportional community composition. This happens after endemism is derived.}

\item{...}{Additional arguments passed to \link[vegan]{vegdist}.}
}
\value{
\code{ps} with a new \code{dissim} element added.
}
\description{
This function calculates pairwise quantitative phylogenetic dissimilarity between communities and returns the
phylospatial object with the dissimilarity data as an element called \code{dissim}.
}
