% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_canape.R
\name{ps_canape}
\alias{ps_canape}
\title{Binary randomization tests including CANAPE}
\usage{
ps_canape(ps, null_model = "curveball", spatial = T, ...)
}
\arguments{
\item{ps}{phylospatial object}

\item{null_model}{see \code{canaper::cpr_rand_test}}

\item{spatial}{Logical: should the function return a spatial object (TRUE, default) or a vector (FALSE).}

\item{...}{further arguments passed to \code{canaper::cpr_rand_test}}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each metric.
}
\description{
This function is a wrapper around the \code{cpr_rand_test} function from the \code{canaper} package.
In keeping with CANAPE, it requires a binary community matrix and uses a hard significance threshold (for an
alternative that utilizes continuous community values and retains a significance gradient, see \code{ps_rand}).
}
\details{
This function runs \code{canaper::cpr_rand_test}; see the help for that function for details.

It also runs \code{canaper::cpr_classify_endem} on the result, and includes the resulting classification as an
additional variable, 'endem_type', in the output. 'endem_type' values 0-4 correspond to not-significant, neo,
paleo, mixed, and super endemesim, respectively.
}
\references{
Mishler, B. D., Knerr, N., González-Orozco, C. E., Thornhill, A. H., Laffan, S. W., & Miller, J. T. (2014).
Phylogenetic measures of biodiversity and neo-and paleo-endemism in Australian Acacia. Nature Communications, 5(1), 4473.

Nitta, J. H., Laffan, S. W., Mishler, B. D., & Iwasaki, W. (2023). canaper: categorical analysis
of neo‐and paleo‐endemism in R. Ecography, 2023(9), e06638.
}
