% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_dissim.R
\name{ps_dissim}
\alias{ps_dissim}
\title{Quantitative phylogenetic dissimilarity}
\usage{
ps_dissim(
  ps,
  method = "bray",
  endemism = FALSE,
  normalize = TRUE,
  add = TRUE,
  ...
)
}
\arguments{
\item{ps}{phylospatial object.}

\item{method}{Character indicating the dissimilarity index to use, passed to \link[vegan]{vegdist}. The default is
"bray", i.e. Bray-Curtis distance, also known as quantitative Sorensen's. See \link[vegan]{vegdist} for a complete
list of options.}

\item{endemism}{Logical indicating whether community values should be divided by column totals (taxon range sizes)
to derive endemism.}

\item{normalize}{Logical indicating whether community values should be divided by row totals (community sums). If \code{TRUE},
dissimilarity is based on proportional community composition. This happens after endemism is derived.}

\item{add}{Logical indicating whether the input phylospatial object should be returned with the dissimilarity matrix
as the \code{dist} element (TRUE, the default) or the dissimilarity matrix should returned alone (FALSE).}

\item{...}{Additional arguments passed to \link[vegan]{vegdist}.}
}
\value{
A pairwise phylogenetic dissimilarity matrix of class \code{dist}, either on its own or as the \code{dissim} element
of \code{sp}.
}
\description{
This function calculates pairwise quantitative phylogenetic dissimilarity between communities.
}
