% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_diversity.R
\name{ps_diversity}
\alias{ps_diversity}
\title{Calculate spatial phylogenetic diversity and endemism metrics}
\usage{
ps_diversity(ps, spatial = T)
}
\arguments{
\item{ps}{phylospatial object (created by \code{phylospatial()} or \code{ps_simulate()}).}

\item{spatial}{Logical: should the function return a spatial object (TRUE, default) or a vector (FALSE).}
}
\value{
A matrix or raster stack with a column or layer (respectively) for each diversity metric.
}
\description{
This function calculates a range of metrics including phylogenetic diversity and endemism, as well as
diversity and endemism of terminals and of "clades" (all terminal and internal taxa in the tree, but
not scaled by branch length). If continuous community data (probabilities or abundances) are provided,
they are used in calculations.
}
\details{
The function calculates the following metrics:
\itemize{
\item TR: Terminal richness, i.e. richness of terminal taxa (in many cases these are species)
\item CR: Clade richness, i.e. richness of taxa at all levels (equivalent to PD on a cladogram)
\item PD: Phylogenetic diversity
\item TE: Terminal endemism, i.e. total endemism-weighted diversity of terminal taxa (a.k.a. "weighted endemism")
\item CE: Clade endemism, i.e. total endemism-weighted diversity of taxa at all levels (equivalent to PE on a cladrogram)
\item PE: Phylogenetic endemism, i.e. endemism-weighted PD
\item Em: Mean endemism (equivalent to CE / CR)
\item RPD: Relative phylogenetic diversity, i.e. branch length of mean resident (equivalent to PD / CR)
\item PEm: Mean phylogenetic endemism, i.e. branch length / range size of mean resident (equivalent to PE / CR)
\item RPE: Relative phylogenetic endemism, i.e. mean endemism-weighted branch length (equivalent to PE / CE)
}
}
