% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps_get_ranges}
\alias{ps_get_ranges}
\title{Get \code{phylospatial} community data in spatial format}
\usage{
ps_get_ranges(ps, tips_only = FALSE)
}
\arguments{
\item{ps}{\code{phylospatial} object.}

\item{tips_only}{Logical indicating whether only the terminal taxa (TRUE) or all taxa (FALSE) should be returned.}
}
\value{
Either a \code{SpatRaster} with a layer for every taxon, or an \code{sf} data frame with a variable for every taxon,
depending on which data type was used to create \code{ps}.
}
\description{
Get \code{phylospatial} community data in spatial format
}
