% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_quantize.R
\name{ps_quantize}
\alias{ps_quantize}
\title{Stratified randomization of a phylospatial object}
\usage{
ps_quantize(ps, wt_row = NULL, wt_col = NULL, ...)
}
\arguments{
\item{ps}{Object of class \code{phylospatial}}

\item{wt_row, wt_col}{Optional square numeric weight matrices controlling which pairs of
rows (sites) or columns (species) are likely to exchange values during randomization.
Enables spatially constrained or functional-group constrained null models; e.g. a
geographic distance decay matrix from \link{ps_geodist} can be transformed and used as
\code{wt_row}. See \link[nullcat]{nullcat} for details. If left unspecified (the default),
gives unweighted randomization.}

\item{...}{Additional arguments passed to \link[nullcat]{quantize}, such as \code{method},
\code{n_strata}, \code{transform}, \code{fixed}, \code{n_iter}, etc.}
}
\value{
A randomized version of \code{ps}
}
\description{
Generates a randomized version of a phylospatial object by extracting the
tip community matrix, permuting it using \code{nullcat::quantize()}, and rebuilding
the phylospatial object using the permuted tip matrix.
}
\details{
The nullcat \link[nullcat]{quantize} routine involves three steps: converting a
quantitative matrix to categorical strata, permuting the resulting categorical
matrix using one of several categorical null model algorithms, and mapping the
randomized categories back to quantitative values. Supply arguments via \code{...}
to control options for each of these stages.

For repeated randomizations to generate a null distribution, it is more efficient
to use \code{ps_rand(fun = "quantize")}, which is structured to avoid unnecessarily
recomputing overhead that is shared across randomizations.
}
\examples{
\donttest{
if (requireNamespace("nullcat", quietly = TRUE)) {
  ps <- ps_simulate(data_type = "prob")
  ps_rand <- ps_quantize(ps, n_strata = 4,
    n_iter = 1000,
    method = "curvecat", fixed = "cell")

  # spatially constrained randomization
  geo <- as.matrix(ps_geodist(ps))
  W <- exp(-geo / median(geo))
  ps_rand <- ps_quantize(ps, n_strata = 4,
    n_iter = 1000,
    method = "curvecat", fixed = "cell",
    wt_row = W)
}
}
}
\seealso{
\code{\link[=ps_rand]{ps_rand()}}, \code{\link[=ps_geodist]{ps_geodist()}}
}
