% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_regions.R
\name{ps_regions}
\alias{ps_regions}
\title{Phyologenetic regionalization}
\usage{
ps_regions(
  ps,
  k = 5,
  method = "kmeans",
  endemism = FALSE,
  normalize = TRUE,
  ...
)
}
\arguments{
\item{ps}{phylospatial object.}

\item{k}{Number of spatial clusters to divide the region into (Positive integer).}

\item{method}{Clustering method. Options include "kmeans", and the methods listed under \link[stats]{hclust}.}

\item{endemism}{Logical indicating whether community matrix values should be divided by column (species) totals.}

\item{normalize}{Logical indicating whether community matrix values should be divided by row (community) totals.
If so, this happens after endemism division.}

\item{...}{Additional arguments passed to \link{ps_dissim}.}
}
\value{
A raster or matrix with an integer indicating which of the \code{k} regions each site belongs to.
}
\description{
Phyologenetic regionalization
}
