% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss.R
\name{moss}
\alias{moss}
\title{Load California moss spatial phylogenetic data}
\source{
Kling, Gonzalez-Ramirez, Carter, Borokini, and Mishler (2024) bioRxiv, https://doi.org/10.1101/2024.12.16.628580.
}
\usage{
moss(format = "raster")
}
\arguments{
\item{format}{Either "raster" (default) or "polygon"}
}
\value{
a \code{phylospatial} object
}
\description{
Get example \code{phylospatial} data set based on a phylogeny and modeled distributions of 443 moss species
across California. This data set is a coarser version of data from Kling et al. (2024). It contains
occurrence probabilities, and is available in raster or polygon spatial formats.
}
\examples{
\donttest{
moss()
}

}
