% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss.R
\docType{data}
\name{moss}
\alias{moss}
\title{California moss spatial phylogenetic data}
\format{
\subsection{\code{moss}}{

A \code{phylospatial} data object
}
}
\source{
This is a spatially coarser version of data published in: Kling, Gonzalez-Ramirez, et al. (2024)
bioRxiv, https://doi.org/10.1101/2024.12.16.628580.
}
\usage{
moss
}
\description{
Phylogeny and modeled distributions of 443 moss species across California. This is a polygon data set
with occurrence probabilities; to get the data in raster format and/or in binary presence-absence data
format, see \link{moss_data}.
}
\keyword{datasets}
