% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss.R
\docType{data}
\name{moss}
\alias{moss}
\title{California moss spatial phylogenetic data}
\format{
\subsection{\code{moss}}{

A \code{phylospatial} data object
}
}
\source{
Kling et al. (2025) Sys. Bot.
}
\usage{
moss
}
\description{
Phylogeny and modeled distributions of 443 moss species across California. This is a polygon data set
with occurrence probabilities; to get the data in raster format and/or in binary presence-absence data
format, see \link{moss_data}.
}
\keyword{datasets}
