% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moss.R
\name{moss_data}
\alias{moss_data}
\title{Load example \code{phylospatial} data on California mosses}
\usage{
moss_data(
  data_type = c("probability", "binary"),
  spatial_type = c("raster", "polygons")
)
}
\arguments{
\item{data_type}{Either "probability" or "binary".}

\item{spatial_type}{Either "raster" or "polygons".}
}
\value{
A \code{phylospatial} object with a phylogeny and modeled distributions of 443 species of mosses,
from Kling et al. (2025) Sys. Bot.
}
\description{
This function returns the \link{moss} \code{phylospatial} data set, with options for raster vs polygon formatting
and for probability vs binary community data type.
}
\examples{
moss_data("prob", "poly")
moss_data("bin", "rast")
}
