% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylospatial-package.R
\docType{package}
\name{phylospatial-package}
\alias{phylospatial-package}
\title{phylospatial: Spatial Phylogenetic Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analyze spatial phylogenetic diversity patterns. Use your data on an evolutionary tree and geographic distributions of the terminal taxa to compute diversity and endemism metrics, test significance with null model randomization, analyze community turnover and biotic regionalization, and perform spatial conservation prioritizations. All functions support quantitative community data in addition to binary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthewkling.github.io/phylospatial/}
  \item \url{https://github.com/matthewkling/phylospatial}
  \item Report bugs at \url{https://github.com/matthewkling/phylospatial/issues}
}

}
\author{
\strong{Maintainer}: Matthew Kling \email{mattkling@berkeley.edu} (\href{https://orcid.org/0000-0001-9073-4240}{ORCID}) [copyright holder]

}
\keyword{internal}
