% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylospatial-package.R
\docType{package}
\name{phylospatial-package}
\alias{phylospatial-package}
\title{phylospatial: Spatial Phylogenetic Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Conduct various spatial phylogenetic analyses, beginning with data on a phylogenetic tree and the geographic distributions of the terminal taxa. Package functionality includes diversity and endemism metrics, turnover and biotic regionalization analyses, null model randomizations, and spatial conservation prioritization. All functions support quantitative community data in addition to binary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthewkling.github.io/phylospatial/}
}

}
\author{
\strong{Maintainer}: Matthew Kling \email{mattkling@berkeley.edu} (\href{https://orcid.org/0000-0001-9073-4240}{ORCID}) [copyright holder]

}
\keyword{internal}
