% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylospatial-package.R
\docType{package}
\name{phylospatial-package}
\alias{phylospatial-package}
\title{phylospatial: Tools for spatial phylogenetic analysis}
\description{
This package computes a variety of spatial phylogenetic metrics, using data on a phylogenetic tree and the geographic distributions of the terminal taxa. Functionality includes diversity and endemism metrics, turnover and biotic regionalization analyses, null model randomizations, and spatial conservation prioritization. All functions support quantitative community data in addition to binary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matthewkling.github.io/phylospatial/}
}

}
\author{
\strong{Maintainer}: Matthew Kling \email{mkling@uvm.edu} (\href{https://orcid.org/0000-0001-9073-4240}{ORCID})

}
\keyword{internal}
