% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylospatial.R
\name{phylospatial}
\alias{phylospatial}
\title{Create a spatial phylogeny object}
\usage{
phylospatial(
  tree,
  comm,
  spatial = NULL,
  data_type = c("auto", "probability", "binary", "abundance"),
  clade_fun = NULL,
  check = TRUE
)
}
\arguments{
\item{tree}{Phylogeny of class \link[ape]{phylo}. Terminals whose names do not match \code{comm} will be dropped with a warning.}

\item{comm}{Community data representing the distribution of terminal taxa across sites. Can be a matrix with a column per terminal and
a row per site, a \link[terra]{SpatRaster} with one layer per terminal, or a \code{sf} data with a column per terminal. Taxa whose names do
not match between column/layer names in \code{comm} and tip labels in \code{tree} will be dropped with a warning.}

\item{spatial}{An optional \code{SpatRaster} layer or \code{sf} object indicating site locations. The number of cells or rows must match \code{comm}.
Ignored if \code{comm} is a \code{SpatRaster} or \code{sf} object.}

\item{data_type}{Character giving the data type of \code{comm}. Must be "binary", "probability", "abundance", or "auto" (the default).
This determines how community values for clades are calculated from the values for terminal taxa. If "binary" (presence-absence),
a clade is considered present in a site if any terminal in the clade is present. If "probability," clade probabilities are calculated
as the probability that at least one terminal is present in a site. If "abundance," clade abundances are calculated as the sum of
abundances for terminals in the clade in a site. If "auto," an attempt is made to guess which of these three data types was provided.
Ignored if \code{comm} already includes clade ranges.}

\item{clade_fun}{Function to calculate the local community weight for a clade based on community weights for tips found in a given location.
Must be either NULL (the default, in which case the default function for the selected \code{data_type} is used) or a summary function that
takes a numeric vector and returns a single numeric output. Ignored if \code{comm} already includes clade ranges.}

\item{check}{Logical indicating whether community data should be validated. Default is TRUE.}
}
\value{
A "phylospatial" object, which is a list containing the following elements:
\itemize{
\item{"data_type"}{Character indicating the community data type}
\item{"tree}{Phylogeny of class \code{phylo}}
\item{"comm"}{Community matrix of community data}
\item{"spatial"}{A \code{SpatRaster} or \code{sf} providing spatial coordinates for the rows in \code{comm}. May be missing if no spatial data was supplied.}
\item{"dissim"}{A community dissimilary matrix of class \code{dist} indicating pairwise phylogenetic dissimilarity between sites. Missing unless
\code{ps_dissim(..., add = T)} is called.}
}
}
\description{
This function creates a \code{phylospatial} object. This is the core data type in the phylospatial library, and
is a required input to most other functions in the package. The two essential components of a spatial phylogeny object
are a phylogenetic tree and an community data set.
}
