% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylospatial.R
\name{plot.phylospatial}
\alias{plot.phylospatial}
\title{Plot a \code{phylospatial} object}
\usage{
\method{plot}{phylospatial}(ps, comp = c("tree", "comm"), max_taxa = 12, ...)
}
\arguments{
\item{ps}{\code{phylospatial} object}

\item{comp}{Either \code{"tree"} or \code{"comm"}, indicating which component to plot.}

\item{max_taxa}{Integer giving the maximum number of taxon ranges to plot, if \code{comp = "tree"}.}

\item{...}{Additional arguments passed to plotting methods.}
}
\description{
Plot a \code{phylospatial} object
}
