% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_prioritize.R
\name{ps_prioritize}
\alias{ps_prioritize}
\title{Phylogenetic conservation prioritization}
\usage{
ps_prioritize(ps, protection = NULL, lambda = 1, level = 1)
}
\arguments{
\item{ps}{phylospatial object.}

\item{protection}{Starting protection status. If this argument is not specified, it is assumed that no existing reserves are present.
Otherwise, must be a numeric vector or \code{SpatRaster} with dimensionality matching the number of sites in \code{ps} and values between
0 and 1 representing the existing level of conservation effectiveness in each site.}

\item{lambda}{Shape parameter for taxon conservation benefit function. This can be any real number. Positive values, such as the default
value \code{1}, place higher priority on conserving the first part of the range of a given species or clade, while negative values
(which are not typically used) place higher priority on fully protecting some taxa than partially protecting all taxa. See the function
\link{lambda} for details.}

\item{level}{Effectiveness level of proposed new reserves (number between 0 and 1, with same meaning as starting \code{protection}).}
}
\value{
A ranking of conservation priorities, with low values representing higher priorities.
}
\description{
Create a ranking of conservation priorities using greedy forward stepwise optimization.
}
