% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_rand.R
\name{ps_rand}
\alias{ps_rand}
\title{Null model randomization analysis of alpha diversity metrics}
\usage{
ps_rand(ps, n_rand = 100, spatial = TRUE, n_cores = 1, progress = TRUE, ...)
}
\arguments{
\item{ps}{\code{phylospatial} object.}

\item{n_rand}{Integer giving the number of random communities to generate.}

\item{spatial}{Logical: should the function return a spatial object (TRUE, default) or a matrix (FALSE).}

\item{n_cores}{Integer giving the number of compute cores to use for parallel processing.}

\item{progress}{Logical: should a progress bar be displayed? Defualt is TRUE.}

\item{...}{Additional arguments passed to \link{quantize}, such as \code{algorithm}, \code{n_iter},
\code{n_strata}, \code{jitter}, \code{transform}, \code{priority}, etc.}
}
\value{
A matrix with a row for every row of \code{x}, a column for every metric in \link{ps_diversity}, and
values indicating the proportion of randomizations in which the observed diversity metric was greater than
the randomized metric.
}
\description{
This function compares to diversity metrics calculated in \link{ps_diversity} to their null distributions
computed by randomizing the community matrix. Randomization is done using the \link{quantize} method for
community matrices containing continuous quantities such as occurrence probabilities or abundances.
}
\examples{
ps <- ps_simulate()
ps_rand(ps)

ps_rand(ps, transform = sqrt, priority = "rows")
}
