% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_regions.R
\name{ps_regions}
\alias{ps_regions}
\title{Phyologenetic regionalization}
\usage{
ps_regions(ps, k = 5, method = "kmeans", endemism = FALSE, normalize = TRUE)
}
\arguments{
\item{ps}{A \code{phylospatial} object. If \code{method} is anything other than \code{"kmeans"}, it must contain a
\code{dissim} component generated by \link{ps_add_dissim}.}

\item{k}{Number of spatial clusters to divide the region into (Positive integer).}

\item{method}{Clustering method. Options include \code{"kmeans"} and any methods listed under \link[stats]{hclust}.}

\item{endemism}{Logical indicating whether community values should be divided by column totals (taxon range sizes)
to derive endemism. Only used if \code{method = "kmeans"}; in other cases this information should instead be
supplied to \link{ps_add_dissim}.}

\item{normalize}{Logical indicating whether community values should be divided by row totals (community sums). If \code{TRUE},
dissimilarity is based on proportional community composition. This happens after endemism is derived. Only used if
\code{method = "kmeans"}; in other cases this information should instead be supplied to \link{ps_add_dissim}.}
}
\value{
A raster or matrix with an integer indicating which of the \code{k} regions each site belongs to.
}
\description{
Phyologenetic regionalization
}
