% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_rgb.R
\name{ps_rgb}
\alias{ps_rgb}
\title{Color space ordination of spatial phylogenetic composition}
\usage{
ps_rgb(ps, method = c("nmds", "cmds", "pca"), trans = identity)
}
\arguments{
\item{ps}{A \code{phylospatial} object with a non-null \code{dissim} component, generated by \link{ps_add_dissim}.}

\item{method}{Ordination method, either "pca" (principal component analysis), "cmds" (classical MDS),
or "nmds" (nonmetric MDS, the default, which is slower but often preferred).}

\item{trans}{A function giving a transformation to apply to each dimension of the ordinated data.
The default is the identity function. Specifying \code{rank} generates a more uniform color distribution.}
}
\value{
A raster or matrix with layers or columns (respectively) containing RGB color values in the range 0-1.
}
\description{
Color space ordination of spatial phylogenetic composition
}
